% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campbell}
\alias{campbell}
\alias{campbell_3d}
\title{campbell atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
campbell

campbell_3d
}
\description{
A historical Campbell atlas based on files
released with  Pijnenburg et al., NeuroImage, 
239, 2021, (\href{https://doi.org/10.1016/j.neuroimage.2021.118274}{DOI})
Version 1; 15-01-2021.
}
\examples{
data(campbell)
data(campbell_3d)
}
\references{
A.W. Campbell; Histological Studies On the Localisation of Cerebral Function; University Press (1905)
    
\itemize{
 \item{campbell}{ - campbell atlas}
 \item{campbell_3d}{ - campbell 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
