% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campbell}
\alias{campbell}
\title{Campbell Cortical Atlas (1905)}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(campbell)
}
\description{
Brain atlas for the Campbell (1905) cortical parcellation
with 17 regions per hemisphere. Contains both 2D polygon geometry
for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(campbell)
campbell
}
\references{
Campbell A.W. (1905). Histological studies on the localisation
of cerebral function. Cambridge University Press.

Pijnenburg R. et al. (2021). Myelo- and cytoarchitectonic microstructural
and functional parcellation of the human cortex based on the Campbell atlas.
\emph{NeuroImage}, 239, 118274. \doi{10.1016/j.neuroimage.2021.118274}
}
\concept{ggseg_atlases}
\keyword{datasets}
